fntx-cli/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
asyncio>=3.11.0
httpx>=0.24.0
pydantic>=2.0.0

# Trading
ccxt>=4.0.0  # Crypto exchanges
ib_insync>=0.9.0  # Interactive Brokers
yfinance>=0.2.0  # Market data

# AI/ML
torch>=2.0.0
stable-baselines3>=2.0.0
transformers>=4.30.0

# Blockchain
web3>=6.0.0
eth-account>=0.9.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# API/Server
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# CLI
click>=8.1.0
rich>=13.0.0
prompt-toolkit>=3.0.0

# MCP
# Note: MCP SDK to be added when available

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0